errors: declare all errors from notmuch statuses.
